requirements/base.txt
Django==4.0.2
django-colorfield==0.6.3
django-environ==0.8.1
django-extensions==3.1.5

Pillow==9.0.1
